requirements.txt
discord.py
asyncio
beautifulsoup4
requests